refactor: YouTubeDataApi 인터페이스 -> 클래스 변경 (#16)
