+++ backend/src/test/java/com/celuveat/celuveat/restaurantlinkfetcher/infra/youtube/YouTubeRestaurantLinkFetcherTest.java
@@ -3,7 +3,7 @@ package com.celuveat.celuveat.restaurantlinkfetcher.infra.youtube;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.celuveat.celuveat.restaurantlinkfetcher.domain.RestaurantLinkFetcher;
-import com.celuveat.celuveat.restaurantlinkfetcher.infra.youtube.api.MockYouTubeDataApiImpl;
+import com.celuveat.celuveat.restaurantlinkfetcher.infra.youtube.api.MockYouTubeDataApi;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
@@ -22,7 +22,7 @@ class YouTubeRestaurantLinkFetcherTest {
     void 한_채널의_모든_영상들의_링크를_반환한다() {
         // given
         RestaurantLinkFetcher restaurantLinkFetcher =
-                new YouTubeRestaurantLinkFetcher(new MockYouTubeDataApiImpl());
+                new YouTubeRestaurantLinkFetcher(new MockYouTubeDataApi());
 
         // when
         List<String> videoIds = restaurantLinkFetcher.fetchAllByChannelId("a");
@@ -36,9 +36,9 @@ class YouTubeRestaurantLinkFetcherTest {
     void 지정한_시간_이후의_모든_영상들의_링크를_반환한다(int month, int day, int expected) {
         // given
         RestaurantLinkFetcher restaurantLinkFetcher =
-                new YouTubeRestaurantLinkFetcher(new MockYouTubeDataApiImpl());
+                new YouTubeRestaurantLinkFetcher(new MockYouTubeDataApi());
         LocalDateTime startDateTime = LocalDateTime.of(2023, month, day, 21, 0, 0);
-        
+
         // when
         List<String> result = restaurantLinkFetcher.fetchNewByChannelId("a", startDateTime);
 
@@ -50,7 +50,7 @@ class YouTubeRestaurantLinkFetcherTest {
     void 지정한_시간이_아주_작으면_모든_영상들의_링크를_반환한다() {
         // given
         RestaurantLinkFetcher restaurantLinkFetcher =
-                new YouTubeRestaurantLinkFetcher(new MockYouTubeDataApiImpl());
+                new YouTubeRestaurantLinkFetcher(new MockYouTubeDataApi());
 
         // when
         List<String> result = restaurantLinkFetcher.fetchNewByChannelId("a", LocalDateTime.MIN);
@@ -63,7 +63,7 @@ class YouTubeRestaurantLinkFetcherTest {
     void 지정한_시간이_아주_크면_아무_영상도_반환하지_않는다() {
         // given
         RestaurantLinkFetcher restaurantLinkFetcher =
-                new YouTubeRestaurantLinkFetcher(new MockYouTubeDataApiImpl());
+                new YouTubeRestaurantLinkFetcher(new MockYouTubeDataApi());
 
         // when
         List<String> result = restaurantLinkFetcher.fetchNewByChannelId("a", LocalDateTime.MAX);
