+++ backend/src/test/java/com/celuveat/celuveat/restaurantlinkfetcher/infra/youtube/api/MockYouTubeDataApiTest.java
@@ -8,15 +8,15 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
 
-@DisplayName("MockYouTubeDataApiImpl 은(는)")
+@DisplayName("MockYouTubeDataApi 은(는)")
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(ReplaceUnderscores.class)
-class MockYouTubeDataApiImplTest {
+class MockYouTubeDataApiTest {
 
     @Test
     void 미리_설정된_데이터를_반환한다() {
         // given
-        YouTubeDataApi youTubeDataApi = new MockYouTubeDataApiImpl();
+        YouTubeDataApi youTubeDataApi = new MockYouTubeDataApi();
 
         // when
         SearchListResponse response = youTubeDataApi.searchList("a");
@@ -28,7 +28,7 @@ class MockYouTubeDataApiImplTest {
     @Test
     void 페이지_토큰으로_다음_응답을_가져온다() {
         // given
-        YouTubeDataApi youTubeDataApi = new MockYouTubeDataApiImpl();
+        YouTubeDataApi youTubeDataApi = new MockYouTubeDataApi();
         SearchListResponse response = youTubeDataApi.searchList("a");
         String nextPageToken = response.nextPageToken();
